Separated tests into resource specific tests and added update test to aws test
